List extracted from veneer.c/.h and object.c/.h
