+++ src/classes/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(morpho
     PRIVATE
         cmplx.c cmplx.h
+        list.c list.h
         range.c range.h
         system.c system.h
 )
